backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
tensorflow-cpu==2.18.0
scikit-learn==1.5.2
numpy==1.26.4
pydantic==2.10.3
python-multipart==0.0.12
